Split implemenation from template declaration
